Move /comparison URL parsing up to Main.
Links to /comparisons when Sidebar didn't exist were resulting in going to comparisonMode with an empty comparisonFeatures. Moving the useEffect up the component hierarchy to a persistent component like Main fixes that.
